requirements/dev.txt
SQLAlchemy==0.8.1
Twisted==13.1.0
amqp==1.0.11
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.28
buildtools==1.0
celery==3.0.19
distribute==0.6.34
eventlet==0.12.1
flower==0.5.1
greenlet==0.4.1
kombu==2.5.10
librabbitmq==1.0.1
mercurial==2.6.3
python-dateutil==2.1
python-ldap==2.4.12
redis==2.7.5
requests==1.2.3
simplejson==3.3.0
six==1.3.0
tornado==3.0.2
wsgiref==0.1.2
zope.interface==4.0.5
